bot/requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
yt-dlp>=2024.0.0
pysrt>=1.1.2
aiohttp>=3.9.0
flask>=3.0.0
